% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_taxonomy_funcs.R
\name{coalesce_manual}
\alias{coalesce_manual}
\title{coalesce_manual: function to manually coalesce rows in attribute table}
\usage{
coalesce_manual(df)
}
\arguments{
\item{df}{dataframe that may need coalescing}
}
\value{
dataframe
}
\description{
coalesce_manual: function to manually coalesce rows in attribute table
}
