% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_manip_functions.R
\name{coalesce_occur}
\alias{coalesce_occur}
\title{Title: function to manually coalesce rows in occurrence table}
\usage{
coalesce_occur(df)
}
\arguments{
\item{df}{dataframe that may need coalescing}
}
\value{
dataframe
}
\description{
Title: function to manually coalesce rows in occurrence table
}
