% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_taxonomy_funcs.R
\name{get_accepted_families}
\alias{get_accepted_families}
\title{get_accepted_families: function to get only accepted family info from GBIF}
\usage{
get_accepted_families(taxa_name)
}
\arguments{
\item{taxa_name}{genus species name of a taxa}
}
\value{
dataframe
}
\description{
get_accepted_families: function to get only accepted family info from GBIF
}
