% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_info_functions.R
\name{get_accepted_taxonomy}
\alias{get_accepted_taxonomy}
\title{Get accepted taxonomic information from GBIF}
\usage{
get_accepted_taxonomy(taxa_name)
}
\arguments{
\item{taxa_name}{list of taxa names to search for in GBIF}
}
\value{
dataframe
}
\description{
If the user supplied taxa is an insect, this function will return the
best accepted match from GBIF.  If the user supplied taxa is not an
insect according to GBIF, it will return "id to non-insect species"
in the genus_species column.  If the user supplied taxa is not found
in GBIF, this function will return "species not found" in the
genus_species column.
}
\examples{
\dontrun{
taxa_list <- c("Abax parallelopipedus", "Hypsicera curvator", "Xylocoris sordidus")
taxa_accepted <- lapply(taxa_list, get_accepted_taxonomy)
}
}
