% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_taxonomy_funcs.R
\name{get_more_info}
\alias{get_more_info}
\title{get_more_info: function to get taxonomic info from other databases besides GBIF}
\usage{
get_more_info(taxa_name)
}
\arguments{
\item{taxa_name}{list of taxa names to search for in other databases}
}
\value{
dataframe
}
\description{
get_more_info: function to get taxonomic info from other databases besides GBIF
}
