% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_taxonomy_funcs.R
\name{separate_attributes}
\alias{separate_attributes}
\title{separate_attributes: function to separate attribute info for each taxa from the raw files}
\usage{
separate_attributes(df_location)
}
\arguments{
\item{df_location}{location of file containing taxa info}
}
\value{
dataframe
}
\description{
separate_attributes: function to separate attribute info for each taxa from the raw files
}
