% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_functions.R
\name{separate_attributes_csv}
\alias{separate_attributes_csv}
\title{separate_attributes_csv: function to separate attribute info for each taxa from the raw files using csv files}
\usage{
separate_attributes_csv(df_location)
}
\arguments{
\item{df_location}{location of file containing taxa info}
}
\value{
dataframe
}
\description{
separate_attributes_csv: function to separate attribute info for each taxa from the raw files using csv files
}
\examples{
\dontrun{
my_df <- "/path/to/my_df.csv"
attrib <- separate_attributes_csv(my_df)
}
}
