% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_functions.R
\name{separate_attributes_xl}
\alias{separate_attributes_xl}
\title{separate_attributes_xl: function to separate attribute info for each taxa from the raw files using xl files}
\usage{
separate_attributes_xl(df_location)
}
\arguments{
\item{df_location}{location of file containing taxa info}
}
\value{
dataframe
}
\description{
separate_attributes_xl: function to separate attribute info for each taxa from the raw files using xl files
}
