% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_taxonomy_funcs.R
\name{separate_occurrence}
\alias{separate_occurrence}
\title{separate_occurrence: function to read in and separate occurrence info}
\usage{
separate_occurrence(df_location)
}
\arguments{
\item{df_location}{location of files containing taxa info}
}
\value{
dataframe
}
\description{
separate_occurrence: function to read in and separate occurrence info
}
