% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_functions.R
\name{separate_occurrence_csv}
\alias{separate_occurrence_csv}
\title{separate_occurrence_csv: function to read in and separate occurrence info using csv files}
\usage{
separate_occurrence_csv(df_location)
}
\arguments{
\item{df_location}{location of files containing taxa info}
}
\value{
dataframe
}
\description{
separate_occurrence_csv: function to read in and separate occurrence info using csv files
}
\examples{
\dontrun{
my_df <- "/path/to/my_df.csv"
attrib <- separate_occurrence_csv(my_df)
}
}
