% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_functions.R
\name{separate_source_csv}
\alias{separate_source_csv}
\title{separate_source_csv: function to read in and separate source info using csv files}
\usage{
separate_source_csv(df_location)
}
\arguments{
\item{df_location}{location of files containing taxa info}
}
\value{
dataframe
}
\description{
separate_source_csv: function to read in and separate source info using csv files
}
\examples{
\dontrun{
my_df <- "/path/to/my_df.csv"
}
}
