% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_taxonomy_funcs.R
\name{separate_taxonomy}
\alias{separate_taxonomy}
\title{separate_taxonomy: a function that cleans the dataframes and separates taxonomy columns}
\usage{
separate_taxonomy(df_location)
}
\arguments{
\item{df_location}{location of files containing raw insect data}
}
\value{
dataframe
}
\description{
separate_taxonomy: a function that cleans the dataframes and separates taxonomy columns
}
