% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_functions.R
\name{separate_taxonomy_csv}
\alias{separate_taxonomy_csv}
\title{separate_taxonomy_csv: a function that cleans the dataframes and separates taxonomy columns using csv files}
\usage{
separate_taxonomy_csv(df_location)
}
\arguments{
\item{df_location}{location of files containing raw insect data}
}
\value{
dataframe
}
\description{
separate_taxonomy_csv: a function that cleans the dataframes and separates taxonomy columns using csv files
}
\examples{
\dontrun{
file_list <- system.file("extdata", "Japan_taxa.xlsx", package = "insectcleanr", mustWork = TRUE)
taxa_list <- lapply(file_list, separate_taxonomy_csv)
}
}
