% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_manip_functions.R
\name{coalesce_manual}
\alias{coalesce_manual}
\title{coalesce_manual: function to manually coalesce rows in attribute table}
\usage{
coalesce_manual(df)
}
\arguments{
\item{df}{dataframe that may need coalescing}
}
\value{
dataframe
}
\description{
coalesce_manual: function to manually coalesce rows in attribute table
}
\examples{
file <- system.file("extdata", "Japan_taxa.xlsx", package = "insectcleanr", mustWork = TRUE)
data_j <- read_excel(file)
data_co <- coalesce_manual(data_j)
}
