% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_info_functions.R
\name{get_accepted_families}
\alias{get_accepted_families}
\title{Get only accepted family-level information from GBIF}
\usage{
get_accepted_families(taxa_name)
}
\arguments{
\item{taxa_name}{genus species name of a taxa}
}
\value{
dataframe
}
\description{
Get only accepted family-level information from GBIF
}
