% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_taxonomy_funcs.R
\name{get_accepted_taxonomy}
\alias{get_accepted_taxonomy}
\title{get_accepted_taxonomy: get only accepted taxonomic info from GBIF}
\usage{
get_accepted_taxonomy(taxa_name)
}
\arguments{
\item{taxa_name}{list of taxa names to search for in GBIF}
}
\value{
dataframe
}
\description{
get_accepted_taxonomy: get only accepted taxonomic info from GBIF
}
